Added AI, Fixed sticky keys

+++ BasicGame/src/com/tutorial/main/BasicEnemy.java
@@ -29,12 +29,12 @@ public class BasicEnemy extends GameObject{
 	
 	public void render(Graphics g) {
 		g.setColor(Color.RED);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 	}
 
 	@Override
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 
 }
