+++ BasicGame/src/com/tutorial/main/Game.java
@@ -14,6 +14,7 @@ import java.util.Random;
  * 4. Enemies, Health Bar
  * 5. Collision, Trail
  * 6. Levels, Spawner, FastEnemy
+ * 7. Sticky Keys, AI
  */
 public class Game extends Canvas implements Runnable{
 
@@ -115,7 +116,7 @@ public class Game extends Canvas implements Runnable{
 		bs.show();
 	}
 	
-	public static int clamp(int var, int min, int max) {
+	public static float clamp(float var, float min, float max) {
 		if(var >= max)
 			return var = max;
 		else if(var <= min)
