+++ BasicGame/src/com/tutorial/main/Trail.java
@@ -16,7 +16,7 @@ public class Trail extends GameObject{
 	
 	private int width, height;
 	
-	public Trail(int x, int y, ID id, int width, int height, float life, Color color, Handler handler) {
+	public Trail(float x, float y, ID id, int width, int height, float life, Color color, Handler handler) {
 		super(x, y, id);
 		this.handler = handler;
 		this.color = color;
@@ -36,7 +36,7 @@ public class Trail extends GameObject{
 		g2d.setComposite(makeTransparent(alpha)); //alpha value
 		
 		g.setColor(color);
-		g.fillRect(x, y, width, height);
+		g.fillRect((int)x, (int)y, width, height);
 		
 		g2d.setComposite(makeTransparent(1)); //solid 1 alpha
 	}
